Edit move CommandLineBuilder and implement Build.
